+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to the z/OS FTP Plug-in for Zowe CLI will be documented in this file.
 
+## Recent Changes
+- Fix `download uss-file` and `view uss-file` the file under the directory of symbol link.
+
 ## `1.8.2`
 - Add some second shortcuts to match with zowe core CLI options.